refactor: 홈 화면 섹션 컴포넌트 분리 및 패키지 구조 개선
